+++ feincms3/pages.py
@@ -83,15 +83,19 @@ class AbstractPage(CTENode):
             nodes[node.id] = node
         return nodes
 
-    def clean(self):
+    def clean_fields(self, exclude=None):
         """
         Check for path uniqueness problems.
         """
+        exclude = [] if exclude is None else exclude
+        super(AbstractPage, self).clean_fields(exclude)
+
         if self.static_path:
             if not self.path:
-                raise ValidationError(_(
-                    'Static paths cannot be empty. Did you mean \'/\'?'
-                ))
+                error = _('Static paths cannot be empty. Did you mean \'/\'?')
+                raise ValidationError(
+                    error if 'path' in exclude else {'path': error}
+                )
         else:
             self.path = '{}{}/'.format(
                 self.parent.path if self.parent else '/',
@@ -108,14 +112,15 @@ class AbstractPage(CTENode):
         )
         for pk, node in self._branch_for_update().items():
             if clash_candidates.filter(path=node.path).exists():
+                error = _(
+                    'The page %(page)s\'s new path %(path)s would'
+                    ' not be unique.'
+                ) % {
+                    'page': node,
+                    'path': node.path,
+                }
                 raise ValidationError(
-                    _(
-                        'The page %(page)s\'s new path %(path)s would'
-                        ' not be unique.'
-                    ) % {
-                        'page': node,
-                        'path': node.path,
-                    },
+                    error if 'path' in exclude else {'path': error}
                 )
 
     def save(self, *args, **kwargs):